Added SAS Token Expiry check during DeviceClient initialization
Fixed DeviceClient class so that, when initialized with a SAS Token, the token is checked for expiration. A SecurityException will be thrown if it has already expired. This check happens within the IotHubConnectionString constructor called by the DeviceClient constructor.

There is also a check for expired tokens after DeviceClient.Open() is called for longer running applications.

Fixed a few tests that were using a fake SAS token that was expired, causing them to fail in undesired ways.