Fixed MQTT overwriting messages in queue when device is inactive and no messageID is present. Replaced underlying map with a queue to stop this.
previously a map was used to store received messages on the device side. This map made it possible for messages to be overwritten erroneously if the messages didn't have a message id or correlation id to uniquely identify them.

This fix changes the underlying data structure to be a thread safe queue instead. That way, identical messages can be sent to the device and nothing is overwritten. It also simplifies the process of reading topic strings and bodies.