Removing incorrectly-added import statement
`import com.sun.org.apache.bcel.internal.generic.RET;` was added to `SendMessagesCommon`, breaking the build when run under JDK 9, as the `RET` class is not exposed as public API in the module. It also turns out that this class is not used in `SendMessagesCommon`, so removing the import does not appear to result in any issue.