fix(iot-dev): Fix issue where MQTT didn't url encode application properties correctly
URLEncoder.Encode fails to encode the character '+', so I've added a manual replace step for this character.

Also removing all character validation for application property keys and values now that this MQTT bug is fixed. There is no longer a device-side reason to validate these values as all protocols are encoding them correctly now such that the protocol doesn't throw an exception. Iot Hub will be the validator, not the client