fix(iot-dev): Fix race condition that caused send/receive task deadlock
Our invoke method tests have been failing lately likely due to a deadlock happening on the device side upon receiving a method invocation.

This commit limits the scope of where the synchronization on send/receive locks takes place to where the wait and notify calls occur. Previoiusly, the queues of the transport layer were in these synchronization blocks, but these queues are already thread safe.

While not necessary, this commit also adds a failsafe wakeup to the send/receive threads every 10 seconds in case other race conditions arise in the future. It's a little wasteful when everything works correctly though.