refactor(iot-dev): Change thread scheduling pattern in deviceIO layer
The difference between scheduleAtFixedRate and scheduleWithFixedDelay is that the former will attempt to queue new threads even when the previous send/receive thread is not done yet. Now that our SDK has these send/receive threads wait to be notified when a message is ready to be sent/received, we don't want to try to schedule new threads every X milliseconds regardless of previous thread completion. We want to schedule a new thread every X milliseconds after the previous thread has completed