feat(security-provider): Add helper function to compute derived symmetric keys, add samples (#1001)
Symmetric key authentication works a bit differently for individual enrollments and enrollment groups. For enrollment groups, users are expected to compute a derived key for each device provisioned under that group. Previously, we didn't expose a function to do that, or have documentation to explain that users have to do it.

This commit splits the symmetric key provisioning samples into individual and group samples so that we can better show the differences in how these flows work. It also exposes a helper function for computing this derived key for convenience.