feat(iot-dev): Fix multiplexing client not reporting unauthorized exceptions
Previously, if a device client registered to the multiplexing client had an out of date connection string, the register call would block infinitely. Now the method will throw an exception instead.

This also adds an optional timeout to the registration/unregistration APIs to allow users to have shorter or longer timeouts than our new default of 1 minute.

Note that this also changes the thrown exceptions of most of the multiplexing client APIs. This is fine since this feature is in preview.