fix(e2e): Add retry logic to e2e fault injection tests for when fault message wasn't sent successfully
Due to some inconsistent behavior from the service, we assign a message expiry time to our fault injection messages so that we don't continuously send the faulting message over and over. The issue is that if the message expires before it is successfully sent, then the test will have failed to fault the connection, and our recovery logic can't be tested.

This change brings some simple retry logic that will continue to send the fault injection message until it reaches the service. This includes a special case for not treating MESSAGE_EXPIRED as a successful send