refactor(ado): Centralize all device and module identity creation and… (#1173)
Once centralized, we can start changing how we provide these identities. We may recycle an identity that was "disposed" earlier for a new test, we may create devices in bulk ahead of time, etc.

Also, this refactors a lot of twin tests to move their test-specific fields into the test instance class, rather than the test class. That eliminates the possibility of a test using a field altered by a different test