fix(iot-dev): clarify how partial successes work for MultiplexingClient bulk registrations, fix bug around partial success case
When registering X devices to an open connection and 1 of them is disabled, for instance, the register call will throw a MultiplexingClientDeviceRegistrationAuthenticationException, but X-1 devices successfully added their session to the active multiplexed connection.

The bug fix here is simply that the transport layer forgot to remove the device from the set of registered devices if it failed to register in a scenario like the above. It was never successfully registered, so the transport layer should have the same state.