fix(iot-dev): clarify/fix how partial successes work for MultiplexingClient bulk registrations (#1231)
When registering X devices to an open connection and 1 of them is disabled, the register call will throw a MultiplexingClientDeviceRegistrationAuthenticationException, but X-1 devices successfully added their session to the active multiplexed connection. This needed to be clarified since the javadocs didn't call this out.

There are two bug fixes here. The first bug fix is simply that the transport layer forgot to remove the device from the set of registered devices if it failed to register in a scenario like the above. It was never successfully registered, so the transport layer should have the same state.

The second bug is that the MultiplexingClient layer should update it's state accordingly in these partial success scenarios as well. Previously, if it threw an exception in a scenario like the above, its local state would show that all X devices are not registered, even when X-1 of them have open sessions on the actual multiplexed connection. I added a few assert statement to an existing test to validate this change.