break(iot-deps): Remove deprecated APIs from deps package
The only exception here is that we had this pattern of deprecated getters that were replaced by a final getter. Since this naming was just a means to avoid breaking changes, I'm proposing we make the original getters final and un deprecate them. These getters are only expected to be used internally in our SDK, so the customer impact would be minimal if any