fix(iot-dev): Fix bug where multiplexingClient throws unexpectedly on close
If a DeviceClient is registered to a multiplexing client and fails to authenticate, the register call will throw as expected. However, the multiplexing client would also throw the same exception later once the multiplexing client is closed. This fix makes it so the multiplexing client doesn't "hold on" to this registration exception and re-report it on close