break(iot-serv, iot-dev): Remove unnecessary synchronization from public method (#1310)
For instance, a given device method client should be allowed to invoke a method on device a while already invoking a method on device b.

These APIs do not share any local resources that we needed to synchronize on. They are just simple HTTP requests to the service that are constructed locally within the method call itself.