fix(iot-dev): Fix bug where multiplexed client saved more session handlers than expected
If a multiplexing client tried to open, failed due to a network issue or something similar, and then later tried to open again, the session handler list would contain 2 session handlers for each multiplexed device instead of the expected 1. The issue was that the catch logic within the open call didn't clear the session handler list in the event of a network issue or interruption.