break(all): Remove iot deps package
All files in the deps package have been moved to the packages that were using them. The unit tests for each deps class has been moved alongside them as well.

In the case where a file in deps was used by multiple packages, the file has been copied to both packages. This was mostly only applicable for the twin classes and a few serializer classes.

There is some impact on users since a few deps classes were exposed in our public API surface. The changes made to our e2e tests and to our samples reflect the entirety of that change, though.