refactor(all): update dependencies, move to dependency management in main pom
The dependency management section in the base pom.xml defines which versions of each library will be used throughout the entire project. This does not mean that each library will take a dependency on each library defined in this dependencyManagement section though. For instance, our device client pom still needs to declare a dependency on our mqtt library. It just doesn't need to specify the version anymore since it is defined in our dependencyManagement in our base pom