refactor(sample): Rework reconnection sample a bit
Previously, the sample wouldn't work well for users who either don't send telemetry/reported properties often or who don't do it at all. This is because the client manager would only attempt to reconnect once it tried to send telemetry/reported properties and failed because the client was closed. This meant that a client could be offline (and missing direct method calls and desired property updates) for a while.

With this rework, the sample will now immediately start reconnecting if a terminal disconnection event occurs regardless of if there is any outgoing telemetry and/or reported properties.

Also replacing System.out.print calls with log.debug calls so that this code can be more easily copied into a user's application.