refactor(iot-service, e2e): Upgrade Guava and Jackson dependencies
Adding a proxy implementation to our code so that we can make some modifications so it is compatible with more recent guava versions. For the most part, this proxy implementation is equivalent to the dependency we previously took

All this proxy code now lives in the "tests/integration/com/microsoft/azure/sdk/iot/helpers/proxy" path.