fix(iot-dev): Fix unnecessary creation of custom symbol instances in amqp layer
Each new symbol instance is saved within proton-j's symbol cache and this cache lives as long as the reactor itself. That means that a multiplexed connection that is never closed and that has infinite devices added and removed over time would cause the symbol cache to grow infinitely.

With this change, the only symbols in the symbol cache will be the constants used as property keys when building the amqp connection/sessions/links like the below:

            this.amqpProperties.put(Symbol.getSymbol("com.microsoft:channel-correlation-id"), deviceId + "/" + moduleId);

#1478