refactor(security-x509, prov-security, iot-dev): Put commons-codec dependency at appropriate level in pom hierarchy
This library is used in the device client package, but is not used in the security provider or the x509 security provider packages.

Since the device client package takes a dependency on the security provider packages, it got this commons-codec package for free previously. Now, it will need to declare an explicit dependency

Builds on #1543 as submitted by @Celebrate-future