Consolidate package version into main pom.xml
The only spots we need to update each release were in the main pom.xml and in these utilities classes. With this change, the utilities classes will just read this version number from the main pom.xml.

With this change, our releases process will only require bumping the versions of each package in one spot, and we can remove some release pipelines